refactor: extract the duplicate request preventor to its own file
