+++ src/index.ts
@@ -3,7 +3,8 @@ import cors from 'cors';
 import { initLogger, fallbackLogger } from '@snapshot-labs/snapshot-sentry';
 import express from 'express';
 import api from './api';
-import { default as rateLimit, duplicateRequestLimit } from './helpers/rateLimit';
+import rateLimit from './helpers/rateLimit';
+import duplicateRequestPreventor from './helpers/duplicateRequestPreventor';
 import shutter from './helpers/shutter';
 import log from './helpers/log';
 import refreshModeration from './helpers/moderation';
@@ -19,7 +20,8 @@ app.disable('x-powered-by');
 app.use(express.json({ limit: '20mb' }));
 app.use(express.urlencoded({ limit: '20mb', extended: false }));
 app.use(cors({ maxAge: 86400 }));
-app.use(rateLimit, duplicateRequestLimit);
+app.use(rateLimit);
+app.use(duplicateRequestPreventor);
 app.set('trust proxy', 1);
 app.use('/', api);
 app.use('/shutter', shutter);
