+++ test/.env.test
@@ -5,5 +5,3 @@ RELAYER_PK=01686849e86499c1860ea0afc97f29c11018cbac049abf843df875c60054076e
 RATE_LIMIT_DATABASE_URL=redis://localhost:6379
 RATE_LIMIT_KEYS_PREFIX=snapshot-sequencer-test:
 NODE_ENV=test
-RATE_LIMIT_DATABASE_URL=redis://localhost:6379
-RATE_LIMIT_KEYS_PREFIX=snapshot-sequencer:
